+++ packages/helium-admin-cli/src/log-hotspot.ts
@@ -6,8 +6,15 @@ import {
   mobileInfoKey,
   rewardableEntityConfigKey,
 } from "@helium/helium-entity-manager-sdk";
+import { lazyDistributorKey, recipientKey, init as initLazy } from "@helium/lazy-distributor-sdk";
 import { daoKey, subDaoKey } from "@helium/helium-sub-daos-sdk";
-import { getAsset, getAssetProof, HNT_MINT, IOT_MINT, MOBILE_MINT } from "@helium/spl-utils";
+import {
+  getAsset,
+  getAssetProof,
+  HNT_MINT,
+  IOT_MINT,
+  MOBILE_MINT,
+} from "@helium/spl-utils";
 import { PublicKey } from "@solana/web3.js";
 import os from "os";
 import yargs from "yargs/yargs";
@@ -32,8 +39,8 @@ export async function run(args: any = process.argv) {
     },
     keySerialization: {
       type: "string",
-      default: "b58"
-    }
+      default: "b58",
+    },
   });
   const argv = await yarg.argv;
   process.env.ANCHOR_WALLET = argv.wallet;
@@ -42,10 +49,17 @@ export async function run(args: any = process.argv) {
 
   const provider = anchor.getProvider() as anchor.AnchorProvider;
   const hemProgram = await initHem(provider);
+  const lazyProgram = await initLazy(provider);
 
-  // @ts-ignore
-  const [keyToAssetK] = keyToAssetKey(daoKey(HNT_MINT)[0], argv.entityKey, argv.keySerialization!);
-  const keyToAsset = await hemProgram.account.keyToAssetV0.fetchNullable(keyToAssetK);
+  const [keyToAssetK] = keyToAssetKey(
+    daoKey(HNT_MINT)[0],
+    argv.entityKey,
+    // @ts-ignore
+    argv.keySerialization!
+  );
+  const keyToAsset = await hemProgram.account.keyToAssetV0.fetchNullable(
+    keyToAssetK
+  );
   console.log("keyToAsset", keyToAssetK.toBase58());
   PublicKey.prototype.toString = PublicKey.prototype.toBase58;
 
@@ -69,7 +83,7 @@ export async function run(args: any = process.argv) {
       await hemProgram.account.iotHotspotInfoV0.fetchNullable(iotInfo)
     );
   } else {
-    console.log("No iot info")
+    console.log("No iot info");
   }
 
   const [info] = await mobileInfoKey(mobileConfigKey, argv.entityKey);
@@ -79,6 +93,20 @@ export async function run(args: any = process.argv) {
       await hemProgram.account.mobileHotspotInfoV0.fetchNullable(info)
     );
   } else {
-    console.log("No mobile info")
+    console.log("No mobile info");
+  }
+
+  if (keyToAsset) {
+    const iotLazy = lazyDistributorKey(IOT_MINT)[0];
+    const mobileLazy = lazyDistributorKey(MOBILE_MINT)[0];
+    const hntLazy = lazyDistributorKey(HNT_MINT)[0];
+
+    const iotRecipient = recipientKey(iotLazy, keyToAsset!.asset);
+    const mobileRecipient = recipientKey(mobileLazy, keyToAsset!.asset);
+    const hntRecipient = recipientKey(hntLazy, keyToAsset!.asset);
+
+    console.log("Iot Recipient", iotRecipient, await lazyProgram.account.recipientV0.fetchNullable(iotRecipient[0]));
+    console.log("Mobile Recipient", mobileRecipient, await lazyProgram.account.recipientV0.fetchNullable(mobileRecipient[0]));
+    console.log("Hnt Recipient", hntRecipient, await lazyProgram.account.recipientV0.fetchNullable(hntRecipient[0]));
   }
 }