+++ programs/lazy-distributor/src/instructions/mod.rs
@@ -3,6 +3,7 @@ pub mod initialize_compression_recipient_v0;
 pub mod initialize_lazy_distributor_v0;
 pub mod initialize_recipient_v0;
 pub mod set_current_rewards_v0;
+pub mod set_current_rewards_v1;
 pub mod update_destination;
 pub mod update_lazy_distributor_v0;
 
@@ -11,5 +12,6 @@ pub use initialize_compression_recipient_v0::*;
 pub use initialize_lazy_distributor_v0::*;
 pub use initialize_recipient_v0::*;
 pub use set_current_rewards_v0::*;
+pub use set_current_rewards_v1::*;
 pub use update_destination::*;
 pub use update_lazy_distributor_v0::*;